- Minor changes in MSP policy. - Fixed bug in MSP availability information for queues without switch values. - Fixed MSP availability information tests - Added scheduler event hooks for statistics. - Corrected slowness computation in LibStarsStatistics - Added slowness computation to the CentralizedScheduler - SimGrid simulator broken, does not compile anymore, changes pending

+++ src/test/GlobalScheduler/AggregationTest.hpp
@@ -174,8 +174,8 @@ public:
         return maxSize;
     }
 
-    double getMeanSize() const {
-        return (double)bytes / messages;
+    unsigned int getMeanSize() const {
+        return bytes / messages;
     }
 
     time_duration getMeanTime() const {
